requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
twilio==7.17.0
python-dotenv==0.19.0
psycopg2-binary==2.9.1
flask-sqlalchemy
flask-migrate